Move bcdofs to SNESContext, add temporary postsolve to zero bc component of update
