+++ firedrake/solving_utils.py
@@ -15,6 +15,49 @@ from functools import cached_property
 from firedrake.formmanipulation import ExtractSubBlock
 from firedrake.logging import warning
 
+from finat.ufl import VectorElement, MixedElement
+
+
+def bcdofs(bc, ghost=True):
+    # Return the global dofs fixed by a DirichletBC
+    # in the numbering given by concatenation of all the
+    # subspaces of a mixed function space
+    Z = bc.function_space()
+    while Z.parent is not None:
+        Z = Z.parent
+
+    indices = bc._indices
+    offset = 0
+
+    for (i, idx) in enumerate(indices):
+        if isinstance(Z.ufl_element(), VectorElement):
+            offset += idx
+            assert i == len(indices)-1  # assert we're at the end of the chain
+            assert Z.sub(idx).block_size == 1
+        elif isinstance(Z.ufl_element(), MixedElement):
+            if ghost:
+                offset += sum(Z.sub(j).dof_count for j in range(idx))
+            else:
+                offset += sum(Z.sub(j).dof_dset.size * Z.sub(j).block_size for j in range(idx))
+        else:
+            raise NotImplementedError("How are you taking a .sub?")
+
+        Z = Z.sub(idx)
+
+    if Z.parent is not None and isinstance(Z.parent.ufl_element(), VectorElement):
+        bs = Z.parent.block_size
+        start = 0
+        stop = 1
+    else:
+        bs = Z.block_size
+        start = 0
+        stop = bs
+    nodes = bc.nodes
+    if not ghost:
+        nodes = nodes[nodes < Z.dof_dset.size]
+
+    return numpy.concatenate([nodes*bs + j for j in range(start, stop)]) + offset
+
 
 def _make_reasons(reasons):
     return {getattr(reasons, r): r
@@ -284,6 +327,18 @@ class _SNESContext(object):
         self._coefficient_mapping = None
         self._transfer_manager = transfer_manager
 
+    @cached_property
+    def bc_iset(self):
+        if self._problem.restrict:
+            return None
+        bcs = self._problem.dirichlet_bcs()
+        V = self._x.function_space()
+        bc_nodes = numpy.unique(numpy.concatenate([bcdofs(bc, ghost=False) for bc in bcs], dtype=PETSc.IntType))
+        bc_nodes = V.dof_dset.lgmap.apply(bc_nodes)
+        bc_is = PETSc.IS().createGeneral(bc_nodes, comm=V.comm)
+        bc_is.sort()
+        return bc_is
+
     def reconstruct(self, problem=None, mat_type=None, pmat_type=None, **kwargs):
         """Reconstruct this _SNESContext instance with new arguments."""
         problem = problem or self._problem
@@ -372,6 +427,9 @@ class _SNESContext(object):
             nullspace._apply(self._pjac, transpose=transpose, near=near)
         if ises is not None:
             nullspace._apply(ises, transpose=transpose, near=near)
+
+    def set_ksp_postsolve(self, snes):
+        snes.getKSP().setPostSolve(self.ksp_postsolve)
 
     @PETSc.Log.EventDecorator()
     def split(self, fields):
@@ -579,6 +637,13 @@ class _SNESContext(object):
         if ctx.Jp is not None:
             assert P.handle == ctx._pjac.petscmat.handle
             ctx._assemble_pjac(ctx._pjac)
+
+    @staticmethod
+    def ksp_postsolve(ksp, rhs, sol):
+        dm = ksp.getDM()
+        ctx = dmhooks.get_appctx(dm)
+        if ctx.pre_apply_bcs and not ctx._problem.restrict:
+            sol.isset(ctx.bc_iset, 0)
 
     @cached_property
     def _assembler_jac(self):
